% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{artImpute}
\alias{artImpute}
\title{Artificially miss and impute each data entry individually by ignoring
outlying values}
\usage{
artImpute(dat, ku = 6, marker.proc.list = NULL, miss.pstat = 0.4,
verbose = FALSE)
}
\arguments{
\item{dat, }{an object of log2-normalized gene (or protein) expressions,
containing markers in rows and samples in columns.}

\item{ku, }{an integer in [1,num.markers], upper bound on the number of
nearest neighbors of a marker}

\item{marker.proc.list, }{the rownames of the data to be processed/imputed.}

\item{miss.pstat, }{the score threshold for ignoring potential outliers
during imputation. miss.pstat = 1 ignores values outside of the box (i.e.,
1st-3rd quartiles).}

\item{verbose, }{logical, to show progress of the algorithm
The algorithm ignores values lying at least (1/miss.pstat)-1 times IQR away
from the box; e.g., use miss.pstat=1 to ignore all values lying outside of
the box; use miss.pstat=0.4 to ignore values lying at least 1.5 x IQR away
from the box; use miss.pstat=0 to employ all data during imputation.}
}
\value{
dat.imp, the imputed data that likely represents the expressions of
the markers in the (matched) normal cohort.
}
\description{
Infers the normal-state expression of a marker based on its
coexpression network, i.e., the weigted average of the marker's nearest
neighbors in the data. The returned imputed data will later be used to
elucidate protruding (dysregulated) events.
}
