% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{clusterData}
\alias{clusterData}
\title{Hierarchical cluster analysis}
\usage{
clusterData(data, annotation_row = NULL, annotation_col = NULL,
annotation_colors = NULL, main = NA, stringency_col = 6,
stringency_row = 4, clustering_distance_rows = "euclidean",
clustering_distance_cols = "euclidean", display_numbers = FALSE,
num_clusters_row = NULL, num_clusters_col = NULL,
cluster_rows = TRUE, cluster_cols = TRUE,
annotate_new_clusters_col = FALSE, zero_white = FALSE,
color_palette = NULL)
}
\arguments{
\item{data}{an object of log2-normalized protein (or gene) expressions,
containing markers in rows and samples in columns.}

\item{annotation_row}{data frame that specifies the annotations shown on
left side of the heat map. Each row defines the features for a specific
row. The rows in the data and in the annotation are matched using
corresponding row names. Note that color schemes takes into account if
variable is continuous or discrete.}

\item{annotation_col}{similar to annotation_row, but for columns.}

\item{annotation_colors}{list for specifying annotation_row and
annotation_col track colors manually. It is possible to define the colors
for only some of the features.}

\item{main}{character string, an overall title for the plot.}

\item{stringency_col}{distinctiveness of the clusters (column) to be
identified by the metafeatures algorithm (proportional to num. of clusters).}

\item{stringency_row}{distinctiveness of the clusters (row) to be
identified by the metafeatures algorithm (proportional to num. of clusters).}

\item{clustering_distance_rows}{distance measure used in clustering rows.
Possible values are "correlation" for Pearson correlation and all the
distances supported by dist, such as "euclidean", etc. If the value is
none of the above it is assumed that a distance matrix is provided.}

\item{clustering_distance_cols}{distance measure used in clustering
columns. Possible values the same as for clustering_distance_rows.}

\item{display_numbers}{logical, determining if the numeric values are also
printed to the cells. If this is a matrix (with same dimensions as original
matrix), the contents of the matrix are shown instead of original values.}

\item{num_clusters_row}{number of clusters the rows are divided into, based
on the hierarchical clustering (using cutree), if rows are not clustered,
the argument is ignored.}

\item{num_clusters_col}{similar to num_clusters_row, but for columns.}

\item{cluster_rows}{logical, determining if the rows should be clustered;
or a hclust object.}

\item{cluster_cols}{similar to cluster_rows, but for columns.}

\item{annotate_new_clusters_col}{logical, to annotate cluster IDs (column)
that will be identified.}

\item{zero_white}{logical, to display 0 values as white in the colormap.}

\item{color_palette}{vector of colors used in heatmap.}
}
\value{
tree, the hierarchical tree structure.

cluster_IDs_row, the (row) cluster identities of the markers.

cluster_IDs_col, the (column) cluster identities of the samples.
}
\description{
Displays the hierarchically clustered data by the "pheatmap"
package.
The numbers of clusters along the markers/samples can be set by the user,
then the cluster structures are estimated by pair-wise analysis.
}
\examples{
set.seed(1)
dat = as.data.frame(matrix(runif(10*10),10,10))
rownames(dat) = paste('marker',1:10,sep='')
colnames(dat) = paste('sample',1:10,sep='')
result = clusterData(dat)
}
