% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{outScores}
\alias{outScores}
\title{Analyze putative outliers}
\usage{
outScores(dat)
}
\arguments{
\item{dat}{an object of log2-normalized protein (or gene) expressions,
containing markers in rows and samples in columns.}
}
\value{
outlier p-statistics
}
\description{
Calculates a statistical measure of each data entry being a
putative outlier
}
\examples{
dat = as.data.frame(matrix(1:(5*10),5,10))
rownames(dat) = paste('marker',1:5,sep='')
colnames(dat) = paste('sample',1:10,sep='')
result = outScores(dat)
}
