% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{rankPerOut}
\alias{rankPerOut}
\title{Rank markers by the percentage of outlying events}
\usage{
rankPerOut(dat.dys, marker.proc.list, dys.sig.thr.upp)
}
\arguments{
\item{dat.dys}{samples' distances to regression line (i.e., dysregulation)
on the scatter plots.}

\item{marker.proc.list}{character array, the rownames of the data to be
processed for outlier analyses.}

\item{dys.sig.thr.upp}{the dysregulation score threshold to elucidate/mark
significantly dysregulated outlier events.}
}
\value{
markers rank-ordered by the percentage of outliers over the samples.

the percentages of outliers corresponding to ranked markers.
}
\description{
Ranks markers in the order of decreasing percentage of outlying
events.
}
