% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{dysReg}
\alias{dysReg}
\title{Analyze dysregulated (protruding) events}
\usage{
dysReg(dat, dat.imp, marker.proc.list = NULL, verbose = FALSE)
}
\arguments{
\item{dat}{an object of log2-normalized protein (or gene) expressions,
containing markers in rows and samples in columns.}

\item{dat.imp}{the imputed data that putatively represents the expressions
of the markers in the (matched) normal states.}

\item{marker.proc.list}{character array, the rownames of the data to be
processed for dysregulation.}

\item{verbose}{logical, to show progress of the algorithm}
}
\value{
samples' distances to regression line (i.e., dysregulation) on the
scatter plots.

the scatter plots.
}
\description{
For each marker processed, draws a scatter plot of matching
values of observed vs imputed expressions.
}
\examples{
dat = as.data.frame(matrix(1:(5*10),5,10))
rownames(dat) = paste('marker',1:5,sep='')
colnames(dat) = paste('sample',1:10,sep='')
dat.imp = artImpute(dat, ku=2)
result = dysReg(dat, dat.imp)
}
