% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{rankPerOut}
\alias{rankPerOut}
\title{Rank markers by the percentage of outlying events}
\usage{
rankPerOut(dat.dys, marker.proc.list, dys.sig.thr.upp)
}
\arguments{
\item{dat.dys}{samples' distances to regression line (i.e., dysregulation)
in the scatter plots.}

\item{marker.proc.list}{the given markers to be processed.}

\item{dys.sig.thr.upp}{the dysregulation score threshold to elucidate/mark
significantly dysregulated outlier events.}
}
\value{
ranked markers are returned in marker.out.exp.per.sor.names with
corresponding percentages in marker.out.exp.per.
}
\description{
Ranks markers in the order of decreasing percentage of outlying
events.
}
