% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{statTest}
\alias{statTest}
\title{Analyze dysregulation significance}
\usage{
statTest(dat, dat.imp, marker.proc.list, pval.insig = 0.2)
}
\arguments{
\item{dat}{an object of log2-normalized gene (or protein) expressions,
containing markers in rows and samples in columns.}

\item{dat.imp}{the imputed data that likely represents the expressions of
the markers in the (matched) normal samples.}

\item{marker.proc.list}{the rownames of the data to be processed/imputed.}

\item{pval.insig}{p-value threshold to determine spurious (null)
dysregulation events.}
}
\value{
dat.imp.test, p-values of the markers significance computed by the
KS test.

dat.imp.test.sig, ranked p-values (KS test) of the significant
markers that are lower than pval.insig.

markers.imp.sig, ranked significantly dysregulated markers with
p-values lower than pval.insig.

dat.imp.test.insig, ranked p-values (KS test) of the insignificant
markers that are greater than pval.insig.

markers.imp.insig, ranked markers exhibiting spurious (null)
dysregulation events with p-values greater than pval.insig.
}
\description{
Rank markers by the significance of deviation of the observed
expressions from the (matched) imputed expressions based on the
Kolmogorov-Smirnov (KS) test.
}
